+++ src/app/App.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 
 import { MainPage, LoginPage, FavoritesPage, OfferPage, NotFoundPage } from '../pages';
-import { Spinner, ErrorModal } from '../components';
-import { PrivateRoute, HistoryRouter } from '../components';
+import { Spinner, ErrorModal, PrivateRoute, HistoryRouter } from '../components';
 
 import { AppRoute } from '../shared';
 import { useAppDispatch, useAppSelector } from '../shared';
